% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download}
\alias{download}
\title{download}
\usage{
download(year = NULL, download = FALSE)
}
\arguments{
\item{year}{\code{integer} since 2018}

\item{download}{\code{logic} TRUE or FALSE,the default value is FALSE}
}
\value{
url
}
\description{
download Catalogue of Life China
}
\details{
Visit the website \url{http://col.especies.cn/download} for more details
}
\examples{
\dontrun{
download(year=2019,download=F)
}
}
\author{
Liuyong Ding
}
