% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/familyID.R
\name{familyID}
\alias{familyID}
\title{using familyID to download family IDs}
\usage{
familyID(query = NULL, apiKey = NULL)
}
\arguments{
\item{query}{\code{string} Family name, or part of family name, supports Latin and Chinese names}

\item{apiKey}{\code{string} You need to apply for the apiKey from \url{http://col.especies.cn/api/document} to run this function}
}
\value{
fids string  Array of family ids
}
\description{
download family IDs via family name, supports Latin and Chinese names
}
\details{
Visit the website \url{http://col.especies.cn/api/document} for more details
}
\examples{
\dontrun{
familyID(query="鳗鲡",apiKey="")

query <- c("鳗鲡","鲤")
fids <- lapply(query,familyID,apiKey="")
(fids <- purrr::transpose(fids))
}
}
\author{
Liuyong Ding
}
