% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_synonyms.R
\name{find_synonyms}
\alias{find_synonyms}
\title{Find synonyms via species name}
\usage{
find_synonyms(query)
}
\arguments{
\item{query}{\code{character} species name,The function is similar to [get_colglobal()].}
}
\value{
object
}
\description{
Find synonyms via species name from Catalogue of Life Global.
}
\details{
Visit the website \url{http://webservice.catalogueoflife.org/col/webservice} for more details.
}
\examples{
\dontrun{
dbentry1 <- get_colglobal(query = "4fdb38d6220462049eab9e3f285144e0", option = "id")
str(dbentry1)
head(dbentry1$results)

dbentry2 <- get_colglobal(query = "Platalea leucorodia", option = "name")
str(dbentry2)
head(dbentry2$results)

find_synonyms("Anguilla anguilla")
}
}
\author{
Liuyong Ding
}
