% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_province.R
\name{get_province}
\alias{get_province}
\title{Statistics on Species or Infraspecies by Province}
\usage{
get_province()
}
\value{
Statistics on species or infraspecies by province.
}
\description{
Download the statistics datasets on species/infraspecies by province at \url{http://sp2000.org.cn/statistics/statistics_map} for more details.
}
\details{
Visit the website \url{http://sp2000.org.cn/statistics/statistics_map} for more details.
}
\examples{
\donttest{
x <- get_province()
head(x)
}
}
\author{
Liuyong Ding
}
