% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_redlistchina.R
\name{get_redlistchina}
\alias{get_redlistchina}
\title{Query Redlist of Chinese Biodiversity}
\format{
assessment status:
\describe{
\item{EX}{Extinct}
\item{EW}{Extinct in the wild}
\item{RE}{Regional Extinct}
\item{CR}{Critically Endangered}
\item{EN}{Endangered}
\item{VU}{Vulnerable}
\item{NT}{Near Threatened}
\item{LC}{Least Concern}
\item{DD}{Data Deficient}
}
}
\usage{
get_redlistchina(
  query = NULL,
  option = NULL,
  taxon = "Amphibians",
  viewDT = FALSE
)
}
\arguments{
\item{query}{\code{string} The string to query for.}

\item{option}{\code{character} There is one required parameter, which is either Chinese Names or Scientific Names. Give eithera Chinese Names or Scientific Names. If an Scientific Names is given, the Chinese Names parameter may not be used. Only exact matches found the name given will be returned. option=c("Chinese Names","Scientific Names").}

\item{taxon}{\code{character} There is one required parameter, taxon=c("Amphibians","Angiospermae","Ascomycetes","Basidiomycetes","Birds","Bryophyta","Gymnospermae","Inland Fishes","Lichens","Mammals","Pteridophyta","Reptiles").}

\item{viewDT}{\code{logic} TRUE or FALSE,the default value is FALSE.}
}
\value{
object
}
\description{
Query Redlist of China’s Biodiversity of Vertebrate, Higher Plants and Macrofungi.
}
\details{
Visit the website \url{http://www.mee.gov.cn} for more details.
}
\examples{
\dontrun{
#query assessment status via Chinese Names or Scientific Names
get_redlistchina(query = "Anguilla", option = "Scientific Names")
get_redlistchina(query = "Anguilla nebulosa", option = "Scientific Names")

#creates an HTML widget to display rectangular data
get_redlistchina(taxon = "Inland Fishes", viewDT = TRUE)
}
}
