% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_df.R
\name{list_df}
\alias{list_df}
\title{Catalogue of Life list(s) convert data frame}
\format{
A data frame with 19 variables:
\describe{
\item{ScientificName}{The scientific name (the accepted name) includes the name and the date of the name}
\item{Synonyms}{Synonyms name, Latin}
\item{ChineseName}{Chinese name}
\item{CommonNames}{Common name}
\item{Kingdom}{Kingdom at taxonTree}
\item{Phylum}{Phylum at taxonTree}
\item{Class}{Class at taxonTree}
\item{Order}{Order at taxonTree}
\item{Family}{Family at taxonTree}
\item{Genus}{Genus at taxonTree}
\item{Species}{Species at taxonTree}
\item{Infraspecies}{Infraspecies at taxonTree}
\item{Distribution}{Distribution of species or infraspecies}
\item{Name}{Full name of reviewer in English or Chinese}
\item{Email}{Organization of the reviewer in  English or Chinese}
\item{Address}{Email address of the reviewer}
\item{Institution}{Address of the reviewer in  English or Chinese}
\item{References}{References}
\item{Download}{Download date}
}
}
\source{
Visit the website \url{http://sp2000.org.cn/api/document} for more details
}
\usage{
list_df(x, db = c("colchina", "colglobal"))
}
\arguments{
\item{x}{\code{list} The result returned by the function [search_checklist] or [get_col_global].}

\item{db}{\code{character} db = c("colchina","colglobal")}
}
\description{
Checklist lists convert data frame.
}
\examples{
\dontrun{
##Set your key
set_search_key <- "your apiKey"

##Search family IDs via family name
familyid <- search_family_id(query = "Anguillidae")

##Search taxon IDs via familyID
taxonid <- search_taxon_id(query = familyid$familyIDs,name = "familyID")

#Download detailed lists via species or infraspecies ID
x1 <- search_checklist(query = taxonid$taxonIDs)
str(x1)
x1 <- list_df(x1,db = "colchina")

#Get Catalogue of Life Global checklist via species name
x2 <- get_col_global(query = c("Anguilla marmorata","Anguilla japonica",
                               "Anguilla bicolor","Anguilla nebulosa",
                               "Anguilla luzonensis"),
                               option = "name")
str(x2)
x2 <- list_df(x2,db = "colglobal")
}
}
\author{
Liuyong Ding
}
