% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_url.R
\name{open_url}
\alias{open_url}
\title{Opening the Catalogue of Life China checklist web pages with query}
\usage{
open_url(query = NULL, name = "accceptedName", language = "en")
}
\arguments{
\item{query}{\code{string} the accceptedName,scientificName,chineseName or commonName.}

\item{name}{\code{string} name = c("accceptedName","scientificName","chineseName","commonName"),the default value is "accceptedName".}

\item{language}{\code{string} currently only two languages are supported: Chinese and English,language=c("en","zh"), the default value is "en".}
}
\description{
Query with the accceptedName,scientificName,chineseName and commonName, returning the web pages that meets the criteria.
}
\details{
Visit the website \url{http://sp2000.org.cn/pageservices/document} for more details.
}
\examples{
\donttest{
open_url(query = "Anguilla marmorata",name = "scientificName",language = 'en')
open_url(query = "Anguilla marmorata",name = "scientificName",language = 'zh')
}
}
\author{
Liuyong Ding
}
