% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_map.R
\name{statistics_map}
\alias{statistics_map}
\title{Statistics on Species or Infraspecies by Province}
\usage{
statistics_map()
}
\value{
Statistics on species or infraspecies by province
}
\description{
download the statistics datasets on species/infraspecies by province at \url{http://col.especies.cn/statistics/statistics_map}
}
\details{
Visit the website \url{http://col.especies.cn/statistics/statistics_map} for more details
}
\examples{
\dontrun{
x <- statistics_map()
head(x)
}
}
\author{
Liuyong Ding
}
