% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonID.R
\name{taxonID}
\alias{taxonID}
\title{using taxonID to download taxon IDs}
\format{query:
\describe{
\item{taxonIDs}{an array of species' ids}
\item{familyID}{family ID, unique value}
\item{scientificName}{the scientific name, or part of the scientific name, supports Latin names and Chinese}
\item{commonName}{common name, or part of common name}
}}
\usage{
taxonID(query = NULL, name = "scientificName", apiKey = NULL)
}
\arguments{
\item{query}{\code{string} familyID ,scientificName or commonName}

\item{name}{\code{string} name=c("familyID","scientificName","commonName"),the default value is "scientificName"}

\item{apiKey}{\code{string} You need to apply for the apiKey from \url{http://col.especies.cn/api/document} to run this function}
}
\description{
download taxon IDs via familyID ,scientificName and commonName
}
\details{
Visit the website \url{http://col.especies.cn/api/document} for more details
}
\examples{
\dontrun{
taxonID(query="鳗鲡",name="commonName",apiKey="")

query <- c("鳗鲡","裂腹鱼")
taxonIDs <- lapply(query,taxonID,name='commonName',apiKey="")
(taxonIDs <- purrr::transpose(taxonIDs))
}
}
\author{
Liuyong Ding
}
