% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoology_dbase_name.R
\name{zoology_dbase_name}
\alias{zoology_dbase_name}
\title{Query details of species in China Animal Scientific Database}
\format{
China Animal Scientific Database
\describe{
\item{1}{Chinese zoology database}
\item{2}{China Animal Map Database}
\item{3}{China Economic Animal Database}
\item{4}{Chinese Bird Database}
\item{5}{Chinese Mammal Database}
\item{6}{China Butterfly Database}
\item{7}{Chinese Bee Database}
\item{8}{China Inland Water Fish Database}
\item{9}{Chinese Amphibian Database}
\item{10}{Chinese Reptile database}
\item{...}{allow additional more databases to be used}
}
}
\usage{
zoology_dbase_name()
}
\value{
A collection of names for all China animal scientific databases
}
\description{
Query the database name and return a collection of names for all databases.
}
\details{
Visit the website \url{http://zoology.especies.cn} for more details.
}
\examples{
\dontrun{
##Set your key
set_search_key("your apiKey",db = "zoology")

#Query China Animal Scientific Database lists
zoology_dbase_name()
}
}
\references{
\url{http://zoology.especies.cn}
}
\author{
Liuyong Ding \email{ly_ding@126.com}

Ke Yang \email{ydyangke@163.com}
}
