% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_key.R
\name{set_search_key}
\alias{set_search_key}
\title{SP2000 API keys}
\usage{
set_search_key(key)
}
\arguments{
\item{key}{\code{string} Value to set apiKey to (i.e. your API key).}
}
\value{
A logical of length one, TRUE is the value was set FALSE if not.
value is returned inside invisible(), i.e. it is not printed to screen
when the function is called.
}
\description{
Apply for the apiKey variable to be used by all search_* functions,
register for \url{http://sp2000.org.cn/api/document} and use an API key. This function allows users to set this key.
}
\examples{
\dontrun{
#Set the apiKey variable to be used by all search_* functions
set_search_key("your apiKey")

}
}
