% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_top20.R
\name{get_top20}
\alias{get_top20}
\title{download the top20 species}
\usage{
get_top20()
}
\value{
top20 species
}
\description{
download the top20 species from \url{http://col.especies.cn}
}
\details{
Visit the website \url{http://col.especies.cn} for more details
}
\examples{
\dontrun{
x <- get_top20()
head(x)
}
}
\author{
Liuyong Ding
}
